feat: refactor PedidoCard and CandidatoCard components for improved structure and functionality
